norm_a {PPRaman}	R Documentation

Normalisation to the Amide I peak

Description

Normalising function using min/max method but focused on the Amide I peak which is set 1.

Usage


norm_a(spectra, bin_size = 1)

Arguments

spectra
Spectra in matrix format for normalising

bin_size
Number of wavenumbers in a bin to shift region to normalise to accordingly

Value

Normalised spectra

[Package PPRaman version 2.0.0 Index]