norm_p {PPRaman}	R Documentation

Normalisation to the Phenylalanine peak

Description

Normalising function using min/max method at phenylalanine peak, setting this to 1 and scaling all other data points comparatively.

Usage


norm_p(spectra, bin_size = 1)

Arguments

spectra
Spectra in matrix format for normalising

bin_size
Number of wavenumbers in a bin to shift region to normalise to accordingly

Value

Normalised spectra

[Package PPRaman version 2.0.0 Index]