norm_snv {PPRaman}	R Documentation

Normalising Spectra using SNV

Description

Normalises spectra using Standard Normal Variate method method which scales data based on the mean and standard deviation.

Usage


norm_snv(spectra)

Arguments

spectra
Spectral like data

Value

Normalised spectra using SNV method

References

Barnes, R. J.; Dhanoa, M. S.; Lister, S. J.Standard Normal Variate Transformation and De-Trending of Near-Infrared Diffuse Reflectance Spectra.Applied Spectroscopy 1989,43, 772–777

[Package PPRaman version 2.0.0 Index]