process_spectra {PPRaman}	R Documentation

Pre-processing spectra with options

Description

Provides a wrapper function to trial different pre-processing methods including binning, smoothing, baseline correction and normalisation. Also contains option for rearranging the order using the pp_order setting.

Usage


process_spectra(
  spectra,
  norm_meth,
  rm_bl,
  RCF_rad,
  poly_bl_order,
  filter_length,
  deriv_order = 0,
  poly_smooth_order,
  smoothing_opt = 0,
  binning_opt = 0,
  bin_size = 1,
  pp_order = "1234"
)

Arguments

spectra
Matrix containing spectra

norm_meth
Choice of normalisation - norm_minmax, norm_vec, norm_snv, norm_a (amide I), norm_p (phenylalanine), none

rm_bl
Choice of baseline removal - rcf (Rolling circle filter), pol (Modified Polynomial), der (derivative), rub (rubberband), none

RCF_rad
Numerical input for RCF radius selection

poly_bl_order
Numerical value = polynomial order for baseline removal

filter_length
Filter length (window) for Savitsky Golay smoothing

deriv_order
Numerical value of derivative order for baseline removal

poly_smooth_order
polynomial order for Savtisky Golay smoothing

smoothing_opt
Binary 1 or 0 option 1 = smoothing on, 0 = smoothing off

binning_opt
Binary 1 or 0 option 1 = binning on, 0 = binning off

bin_size
Numeric value for the desired # of wavenumbers in each bin

pp_order
1 = Binning (1 must always come first), 2 = smoothing parameters, 3 = baseline correction, 4 = normalisation

Value

Preprocessed spectra

References

CHAD A. LIEBER and ANITA MAHADEVAN-JANSEN: Automated Method for Subtraction of Fluorescence from Biological Raman Spectra

Brandt, N. N.;Brovko, O. O.;Chiki-shev, A. Y.; Paraschuk, O. D. Optimization of the Rolling-Circle Filter for Raman Back-ground Subtraction. Applied Spectroscopy2006,60, 288–293.

Savitzky, A.; Golay, M. J. E. Smoothing and Differentiation of Data by Simplified LeastSquares Procedures.1964,36, 1627–1639.

[Package PPRaman version 2.0.0 Index]