read_in_shift {PPRaman}	R Documentation

Shift Interp Function

Description

Read in all spectra from file directory, shift and interpolate to 1004 peak (phenylalanine) to adjust for any difference in raw spectra measurement. This function is set up to handle ID's and/or flags in filenames.

Usage


read_in_shift(file_dir)

Arguments

file_dir
Full file directory to shift

Value

Interpolated and Shifted Spectra to 1004

[Package PPRaman version 2.0.0 Index]