read_in_shift_general {PPRaman}	R Documentation

Read in spectra, shift and interpolate

Description

Reads in spectra with x in first column, y in second, shifts to common reference peak and interpolate onto common wavenumber axis. Recommended use with list of spectra read in from .txt files using lapply.

Usage


read_in_shift_general(
  spectra,
  abscissa,
  reference_peak = 1004,
  reference_index = c(300:380)
)

Arguments

spectra
Spectra with first column = original x axis, y = intensity.

abscissa
x-axis to interpolate onto.

reference_peak
Peak to shift spectra to, default = 1004 for phenylalanine in blood serum.

reference_index
Index range for finding maximum intensity around reference peak, default = 300:380 for phenylalanine

Value

Shifted and Interpolated Spectra

[Package PPRaman version 2.0.0 Index]